adiciona aulas 09.2 & 10.1

+++ aula08.2/exemplo/ifClick/ifClick.pde
@@ -1,4 +1,4 @@
-float tamanho;
+int tamanho;
 
 void setup()
 {
